requirements.txt
torch>=1.13.0
transformers>=4.31.0
scikit-learn>=1.2.2
pandas>=1.5.3
numpy>=1.23.5
shutil
python-dotenv
spacy
textblob
fastapi
uvicorn
torch
safetensors
